requirements.txt
numpy>=2.3.5
scipy>=1.16.3
matplotlib>=3.10.7
Pillow>=12.0.0
